change chatMoveToTrash request msgid toInt to toLong

+++ features/user/topchat/src/main/java/com/tokopedia/topchat/chatlist/fragment/ChatListFragment.kt
@@ -519,7 +519,7 @@ open class ChatListFragment constructor() : BaseListFragment<Visitable<*>, BaseA
     }
 
     override fun deleteChat(element: ItemChatListPojo, itemPosition: Int) {
-        chatItemListViewModel.chatMoveToTrash(element.msgId.toInt())
+        chatItemListViewModel.chatMoveToTrash(element.msgId)
         itemPositionLongClicked = itemPosition
     }
